refactor(calc): damage-calculator を責務別に分割
490 行に膨らんでいた damage-calculator.ts を、責務別の 5 モジュールに
切り出し、Facade クラスから組み立てる構造に変更する。公開 API は維持。

- calc/types.ts: 型定義集約 (PokemonInput / DamageCalcResult 等)
- calc/resolvers/resolve-name.ts: 名前解決ヘルパー
- calc/builders/pokemon-builder.ts: Pokemon コンストラクタ引数ビルダー
  (buildPokemonOptions / buildSpeciesOverrides)
- calc/builders/field-builder.ts: Field コンストラクタ引数ビルダー
- calc/formatters/result-formatter.ts: flattenDamage / toPercent

damage-calculator.ts は Facade として 321 行に減量 (-35%)。
呼び出し側 (tools 配下) は public API 経由の import で無変更。
Pokemon HP ダメージ→% 変換の重複 3 箇所を toPercent に統合。

検証: 262 テスト全通過、ビルド成功。
